requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dominate==2.9.1
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
fonttools==4.58.5
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mysqlclient==2.2.7
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
